refactor: split client facade into focused modules
